seperated cp-assets to run parallel, got regression: fa url replacement does not work
